backend/requirements.txt
fastapi==0.115.5
openai==1.54.4
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.2.0
